chore(mme): Split s1ap_mme_handlers_test into multiple targets
